一部機能修正
